fix: typescript-eslint ^5.36.0 (#157)
To be packaged in the next skuba release to ensure these transitive
dependency are upgraded with TypeScript 4.8 support and to avoid this
ugly warning:

```
=============

WARNING: You are currently running a version of TypeScript which is not officially supported by @typescript-eslint/typescript-estree.

You may find that it works just fine, or you may not.

SUPPORTED TYPESCRIPT VERSIONS: >=3.3.1 <4.8.0

YOUR TYPESCRIPT VERSION: 4.8.2

Please only submit bug reports when using the officially supported version.

=============
```

- https://github.com/typescript-eslint/typescript-eslint/releases/tag/v5.36.0
- seek-oss/skuba#954